application/RAG/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.3.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
langdetect==1.0.9
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
weaviate==0.1.2
weaviate-client==4.10.4
